+++ Logger/src/main/kotlin/net/milosvasic/logger/Logger.kt
@@ -2,6 +2,15 @@ package net.milosvasic.logger
 
 interface Logger {
 
+    companion object {
+
+        fun getMessage(e: Exception) = if (e.message == null) {
+            "Exception without message has been caught: $e"
+        } else {
+            e.message as String
+        }
+    }
+
     fun v(tag: String, message: String)
 
     fun d(tag: String, message: String)
@@ -15,4 +24,6 @@ interface Logger {
     fun w(tag: String, message: String)
 
     fun e(tag: String, message: String)
+
+    fun e(tag: String, exception: Exception)
 }